requirements.txt
# AI Trading Bot v2 - Requirements
# Python 3.11+

# === 핵심 의존성 ===
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# === 데이터베이스 ===
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# === 캐싱/메시지 ===
redis>=5.0.0
aioredis>=2.0.0

# === 데이터 분석 ===
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0

# === 기술적 지표 ===
ta-lib>=0.4.28
pandas-ta>=0.3.14b

# === 시각화 (선택) ===
matplotlib>=3.8.0
plotly>=5.18.0

# === HTTP/WebSocket ===
websockets>=12.0
httpx>=0.25.0
requests>=2.31.0

# === 스케줄링 ===
apscheduler>=3.10.0

# === 로깅 ===
loguru>=0.7.0
rich>=13.7.0

# === 설정 ===
pyyaml>=6.0.0
toml>=0.10.0

# === 테스트 ===
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# === ML (선택적) ===
scikit-learn>=1.3.0
xgboost>=2.0.0

# === 텔레그램 알림 ===
python-telegram-bot>=20.7

# === 뉴스/LLM ===
openai>=1.6.0
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# === 유틸리티 ===
tenacity>=8.2.0
python-dateutil>=2.8.0
pytz>=2023.3
